--- packages/transitions/src/zoom-blur-overlay.tsx
@@ -1,241 +1,0 @@
-import type {RefObject} from 'react';
-import React, {useLayoutEffect, useRef} from 'react';
-import {AbsoluteFill, useVideoConfig} from 'remotion';
-
-const VERTEX_SHADER = `#version 300 es
-in vec2 a_pos;
-out vec2 v_uv;
-void main() {
-	v_uv = a_pos * 0.5 + 0.5;
-	gl_Position = vec4(a_pos, 0.0, 1.0);
-}`;
-
-const FRAGMENT_SHADER = `#version 300 es
-precision highp float;
-
-uniform sampler2D u_prev;
-uniform sampler2D u_next;
-uniform float u_time;
-
-in vec2 v_uv;
-out vec4 outColor;
-
-const int SAMPLES = 16;
-const float STRENGTH = 0.35;
-
-vec4 zoomBlur(sampler2D tex, vec2 uv, float strength) {
-	vec2 dir = uv - 0.5;
-	vec4 acc = vec4(0.0);
-	for (int i = 0; i < SAMPLES; i++) {
-		float t = float(i) / float(SAMPLES - 1);
-		float scale = 1.0 - strength * t;
-		acc += texture(tex, 0.5 + dir * scale);
-	}
-	return acc / float(SAMPLES);
-}
-
-void main() {
-	float mixT = fract(u_time / 60.0);
-	vec4 prev = zoomBlur(u_prev, v_uv, STRENGTH * mixT);
-	vec4 next = zoomBlur(u_next, v_uv, STRENGTH * (1.0 - mixT));
-	outColor = mix(prev, next, mixT);
-}`;
-
-const compileShader = (
-	gl: WebGL2RenderingContext,
-	source: string,
-	type: number,
-): WebGLShader => {
-	const shader = gl.createShader(type);
-	if (!shader) {
-		throw new Error('Failed to create shader');
-	}
-
-	gl.shaderSource(shader, source);
-	gl.compileShader(shader);
-	if (!gl.getShaderParameter(shader, gl.COMPILE_STATUS)) {
-		const log = gl.getShaderInfoLog(shader);
-		gl.deleteShader(shader);
-		throw new Error(`Failed to compile shader: ${log}`);
-	}
-
-	return shader;
-};
-
-const createProgram = (gl: WebGL2RenderingContext): WebGLProgram => {
-	const program = gl.createProgram();
-	if (!program) {
-		throw new Error('Failed to create WebGL program');
-	}
-
-	const vs = compileShader(gl, VERTEX_SHADER, gl.VERTEX_SHADER);
-	const fs = compileShader(gl, FRAGMENT_SHADER, gl.FRAGMENT_SHADER);
-	gl.attachShader(program, vs);
-	gl.attachShader(program, fs);
-	gl.linkProgram(program);
-	if (!gl.getProgramParameter(program, gl.LINK_STATUS)) {
-		const log = gl.getProgramInfoLog(program);
-		gl.deleteProgram(program);
-		throw new Error(`Failed to link program: ${log}`);
-	}
-
-	return program;
-};
-
-const createTexture = (gl: WebGL2RenderingContext): WebGLTexture => {
-	const tex = gl.createTexture();
-	if (!tex) {
-		throw new Error('Failed to create texture');
-	}
-
-	gl.bindTexture(gl.TEXTURE_2D, tex);
-	gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_WRAP_S, gl.CLAMP_TO_EDGE);
-	gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_WRAP_T, gl.CLAMP_TO_EDGE);
-	gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_MIN_FILTER, gl.LINEAR);
-	gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_MAG_FILTER, gl.LINEAR);
-	gl.texImage2D(
-		gl.TEXTURE_2D,
-		0,
-		gl.RGBA,
-		1,
-		1,
-		0,
-		gl.RGBA,
-		gl.UNSIGNED_BYTE,
-		new Uint8Array([0, 0, 0, 0]),
-	);
-	return tex;
-};
-
-type GLState = {
-	gl: WebGL2RenderingContext;
-	program: WebGLProgram;
-	prevTex: WebGLTexture;
-	nextTex: WebGLTexture;
-	uTime: WebGLUniformLocation | null;
-	uPrev: WebGLUniformLocation | null;
-	uNext: WebGLUniformLocation | null;
-};
-
-type Props = {
-	readonly prevImageRef: RefObject<ElementImage | null>;
-	readonly nextImageRef: RefObject<ElementImage | null>;
-};
-
-export const ZoomBlurOverlay: React.FC<Props> = ({
-	prevImageRef,
-	nextImageRef,
-}) => {
-	const {width, height} = useVideoConfig();
-	const canvasRef = useRef<HTMLCanvasElement>(null);
-	const stateRef = useRef<GLState | null>(null);
-	const timeRef = useRef(0);
-
-	useLayoutEffect(() => {
-		const canvas = canvasRef.current;
-		if (!canvas) {
-			return;
-		}
-
-		const gl = canvas.getContext('webgl2', {premultipliedAlpha: true});
-		if (!gl) {
-			return;
-		}
-
-		const program = createProgram(gl);
-		const prevTex = createTexture(gl);
-		const nextTex = createTexture(gl);
-
-		const vao = gl.createVertexArray();
-		gl.bindVertexArray(vao);
-		const buffer = gl.createBuffer();
-		gl.bindBuffer(gl.ARRAY_BUFFER, buffer);
-		gl.bufferData(
-			gl.ARRAY_BUFFER,
-			new Float32Array([-1, -1, 1, -1, -1, 1, 1, 1]),
-			gl.STATIC_DRAW,
-		);
-		const aPos = gl.getAttribLocation(program, 'a_pos');
-		gl.enableVertexAttribArray(aPos);
-		gl.vertexAttribPointer(aPos, 2, gl.FLOAT, false, 0, 0);
-
-		stateRef.current = {
-			gl,
-			program,
-			prevTex,
-			nextTex,
-			uTime: gl.getUniformLocation(program, 'u_time'),
-			uPrev: gl.getUniformLocation(program, 'u_prev'),
-			uNext: gl.getUniformLocation(program, 'u_next'),
-		};
-
-		return () => {
-			gl.deleteProgram(program);
-			gl.deleteTexture(prevTex);
-			gl.deleteTexture(nextTex);
-			stateRef.current = null;
-		};
-	}, []);
-
-	useLayoutEffect(() => {
-		const state = stateRef.current;
-		if (!state) {
-			return;
-		}
-
-		const {gl, program, prevTex, nextTex, uTime, uPrev, uNext} = state;
-		const prevImage = prevImageRef.current;
-		const nextImage = nextImageRef.current;
-
-		gl.viewport(0, 0, width, height);
-		gl.clearColor(0, 0, 0, 0);
-		gl.clear(gl.COLOR_BUFFER_BIT);
-		gl.useProgram(program);
-
-		gl.activeTexture(gl.TEXTURE0);
-		gl.bindTexture(gl.TEXTURE_2D, prevTex);
-		if (prevImage) {
-			gl.texElementImage2D(
-				gl.TEXTURE_2D,
-				0,
-				gl.RGBA,
-				gl.RGBA,
-				gl.UNSIGNED_BYTE,
-				prevImage,
-			);
-		}
-
-		gl.uniform1i(uPrev, 0);
-
-		gl.activeTexture(gl.TEXTURE1);
-		gl.bindTexture(gl.TEXTURE_2D, nextTex);
-		if (nextImage) {
-			gl.texElementImage2D(
-				gl.TEXTURE_2D,
-				0,
-				gl.RGBA,
-				gl.RGBA,
-				gl.UNSIGNED_BYTE,
-				nextImage,
-			);
-		}
-
-		gl.uniform1i(uNext, 1);
-
-		gl.uniform1f(uTime, timeRef.current);
-		timeRef.current += 1;
-
-		gl.drawArrays(gl.TRIANGLE_STRIP, 0, 4);
-	});
-
-	return (
-		<AbsoluteFill style={{pointerEvents: 'none'}}>
-			<canvas
-				ref={canvasRef}
-				width={width}
-				height={height}
-				style={{width: '100%', height: '100%'}}
-			/>
-		</AbsoluteFill>
-	);
-};
